Add pre-commit
This commit adds [pre-commit](https://pre-commit.com/) to the projects
generated from this cookiecutter project and configures the hooks that
it should use. These hooks check to ensure that we don't accidentally
commit secrets, as well as checking to ensure that the outputs of any
Jupyter notebooks are removed.

A `requirements-dev.txt` file is also added which contains the
dependencies for pre-commit.

(cherry picked from commit f73a2491a12ba93675db81ce50ecb4de7178f9e2)