Bugfix data folders not being kept in cookiecutter output
The local cookiecutter output contains the `data` folder, and its
sub-folders.

However, these folders are not pushed up to the remote, and therefore
are not accessible to other contributors to the repository.

Based on https://stackoverflow.com/a/20652768, have amended the
`.gitignore` file to keep only the `data/README.md` and the `.gitkeep`
files in the `data` sub-folders, and to ignore everything else in these
folders.

Based on PR #1 comments here:
https://github.com/ukgovdatascience/govcookiecutter/pull/1#discussion_r413866866