Show or hide R instructions based on user-input
This is so that when the user selects that they do not want R pre-commit hooks, then they will not get instructions for installing these hooks on their `README.md`.

Also remove fragments of R code in parent directory. This is because we do not want to impose additional R constraint for contributors of the project.