Add sed command to .envrc to create a .env file
Now, every time a user runs direnv allow, a .env file containing all
the environment variables will be created. This is useful, for example,
to load environment variables for PyCharm via the EnvFile plugin.

Also fix test_envrc.py to now allow for relative paths as well, and
remove functionality to check for self-referencing .envrc environment
variables, as this is no longer supported.