Remove the post-generation hook that creates a `.secrets` file
Also add a `.secrets` file in the template folder, as the hook is no
longer required.

Note that, although we are version controlling `.secrets` here, it will
not be version controlled in any generated folder structure. This is
because using this cookiecutter does not initialise a git repository in
the generated folder structure, so the first time this is done the
`.gitignore` will exclude the `.secrets` file.